new structure analysis.py and plots.py
